setting up for writing refactoring
