+++ src/ws2st/common.hpp
@@ -1,6 +1,7 @@
 #pragma once
 
 #include "clang/Frontend/FrontendActions.h"
+#include "clang/Tooling/Refactoring.h"
 #include "clang/Tooling/CommonOptionsParser.h"
 #include "clang/Tooling/Tooling.h"
