+++ src/ws2st/refactor.hpp
@@ -1,0 +1,19 @@
+/**
+ * @file refactor.hpp
+ * @author james
+ * @brief Idea is to write out the results of the refactoring using the classes defined here
+ * @version 0.1
+ * @date 2024-01-22
+ * 
+ * @copyright Copyright (c) 2024
+ * 
+ */
+#include "common.hpp"
+
+namespace wash {
+
+    class EnumRefactoring : public RefactoringTool {
+        
+    };
+
+}
